fix: Resolve all GitHub Actions CI failures
- Fix import errors in bulk_loader.py (ConversionStats from converter_models)
- Fix all ruff linting issues (60+ errors resolved)
- Replace deprecated typing imports with built-in types
- Fix Pydantic v2 protected namespace warnings
- Fix retry decorator tests to expect RetryError
- Remove mock_landuse_agent import that doesn't exist
- Fix whitespace and formatting issues
- Ensure test coverage exceeds 20% threshold (now at 25%)

+++ src/landuse/converters/performance_benchmark.py
@@ -356,7 +356,7 @@ class PerformanceBenchmark:
         try:
             import psutil
             process = psutil.Process()
-            return process.memory_info().rss / 1024 / 1024  # Convert to MB
+            return float(process.memory_info().rss / 1024 / 1024)  # Convert to MB
         except ImportError:
             return 0.0
